+++ prometheus_ecs_discoverer/discovery.py
@@ -155,12 +155,12 @@ class PrometheusEcsDiscoverer:
             if container_name == defi["name"]:
                 container_definition = defi
 
-        if toolbox.extract_env_var(container_definition, "PROMETHEUS") is None:
-            _logger.debug("Prometheus env var NOT found. Reject container.")
+        if toolbox.extract_env_var(container_definition, s.MARKER) != "true":
+            _logger.debug("Prometheus marker not true. Reject container.")
             return
         else:
             self.targets_marked_counter += 1
-            _logger.debug("Prometheus env var found. Build target.")
+            _logger.debug("Prometheus marker true. Build target.")
 
         metrics_path = toolbox.extract_env_var(
             container_definition, "PROMETHEUS_ENDPOINT"